requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
diffusers==0.36.0
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
setuptools==80.10.2
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
